feat::likeCount

+++ src/like/like.service.ts
@@ -11,23 +11,40 @@ export class LikeService {
     @InjectRepository(Like)
     private likeRepository: Repository<Like>,
     @InjectRepository(Board)
-    private postRepository: Repository<Board>,
+    private boardRepository: Repository<Board>,
   ) {}
 
-  async likePost(postId: number, user: User): Promise<Board> {
-    const post = await this.postRepository.findOne({ where: { id: postId } });
+  async likePost(postId: number, user: User): Promise<void> {
+    const board = await this.boardRepository.findOne({ where: { id: postId } });
+
+    if (!Board) {
+      throw new Error('Post not found');
+    }
+
     const like = new Like();
-    like.board = post;
+    like.board = board;
     like.user = user;
     await this.likeRepository.save(like);
-    return (await this.postRepository.findOne({where:{id:postId}}));
+
+    board.likesCount += 1;
+    await this.boardRepository.save(board);
   }
 
-  async unlikePost(postId: number, user: User): Promise<Board> {
+  async unlikePost(postId: number, user: User): Promise<void> {
     const like = await this.likeRepository.findOne({
       where: { board: { id: postId }, user: { id: user.id } },
     });
+
+    if (!like) {
+      throw new Error('Like not found');
+    }
+
     await this.likeRepository.remove(like);
-    return await this.postRepository.findOne({ where: { id: postId } });
+
+    const post = await this.boardRepository.findOne({ where: { id: postId } });
+    if (post) {
+      post.likesCount -= 1;
+      await this.boardRepository.save(post);
+    }
   }
 }
